Add scripts and apges to handle FAQs and tutorials:
FAQs and tutorials are located at /FAQ/ and /FAQ/tutorials/. The difference between FAQs and tutorials is simple, but important for the implementation: Tutorials are mostly long instructions with many images. FAQs answer with a short text. 
FAQs and tutorials are bundled via three categories: player, download, bugs .
Every tutorials has a single page and FAQs and links to tutorials with the same category are displayed on one page.