Add support for referencing FAQs and tutorials via a hash in the URL
The specified FAQ entry will be expanded and the window scrolls to it automatically.
Add an anchor icon on the right side of the FAQ and tutorial tiles to copy the URL directly. Clicking on the anchor copies either the URL to the tutorial page or the URL of the current page referencing the FAQ entry.